requirements.txt
tabulate>=0.8.10
pytest>=7.4.4
python-dotenv>=1.0.0